Rename ChatMessageEvent to ChatReceivedEvent
