+++ src/main/java/me/dreamhopping/pml/mods/TestMod.java
@@ -2,7 +2,7 @@ package me.dreamhopping.pml.mods;
 
 import me.dreamhopping.pml.events.EventBus;
 import me.dreamhopping.pml.events.InvokeEvent;
-import me.dreamhopping.pml.events.core.client.ChatMessageEvent;
+import me.dreamhopping.pml.events.core.client.chat.ChatReceivedEvent;
 import me.dreamhopping.pml.events.core.mod.ModInitEvent;
 import me.dreamhopping.pml.mods.core.Mod;
 import org.apache.logging.log4j.LogManager;
@@ -22,7 +22,7 @@ public class TestMod {
     }
 
     @InvokeEvent
-    public void onChatMessage(ChatMessageEvent event) {
+    public void onChatMessage(ChatReceivedEvent event) {
         logger.info("TestMod received a chat message: " + event.message);
     }
 }
