+++ src/mc1.14.4/java/me/dreamhopping/pml/transformers/impl/InGameHudTransformerImpl.java
@@ -1,7 +1,7 @@
 package me.dreamhopping.pml.transformers.impl;
 
 import me.dreamhopping.pml.events.EventBus;
-import me.dreamhopping.pml.events.core.client.ChatMessageEvent;
+import me.dreamhopping.pml.events.core.client.chat.ChatReceivedEvent;
 import me.dreamhopping.pml.transformers.InGameHudTransformer;
 import net.minecraft.text.Text;
 
@@ -12,6 +12,6 @@ import net.minecraft.text.Text;
 public class InGameHudTransformerImpl {
     public static void addChatMessage(Text text) {
         // Post the chat message event
-        EventBus.INSTANCE.post(new ChatMessageEvent(text.getString()));
+        EventBus.INSTANCE.post(new ChatReceivedEvent(text.getString()));
     }
 }
