+++ src/mc1.9.4/java/me/dreamhopping/pml/transformers/impl/GuiNewChatTransformerImpl.java
@@ -1,7 +1,7 @@
 package me.dreamhopping.pml.transformers.impl;
 
 import me.dreamhopping.pml.events.EventBus;
-import me.dreamhopping.pml.events.core.client.ChatMessageEvent;
+import me.dreamhopping.pml.events.core.client.chat.ChatReceivedEvent;
 import net.minecraft.util.text.ITextComponent;
 
 /**
@@ -11,6 +11,6 @@ import net.minecraft.util.text.ITextComponent;
 public class GuiNewChatTransformerImpl {
     public static void printChatMessage(ITextComponent chatComponent) {
         // Post the chat message event
-        EventBus.INSTANCE.post(new ChatMessageEvent(chatComponent.getUnformattedText()));
+        EventBus.INSTANCE.post(new ChatReceivedEvent(chatComponent.getUnformattedText()));
     }
 }
